+++ domain/src/main/kotlin/com/astrainteractive/astrarating/domain/entities/tables/dto/UserRatingDTO.kt
@@ -3,9 +3,9 @@ package com.astrainteractive.astrarating.domain.entities.tables.dto
 import com.astrainteractive.astrarating.domain.api.NON_EXISTS_KEY
 
 data class UserRatingDTO(
-    val id: Long = NON_EXISTS_KEY,
-    val userCreatedReport: Long,
-    val reportedUser: Long,
+    val id: Int = NON_EXISTS_KEY,
+    val userCreatedReport: Int,
+    val reportedUser: Int,
     val rating: Int,
     val message: String,
     val time: Long = System.currentTimeMillis(),
